Reflect verbosity option changes for Makefile (#1000)
Prior to this patch, -q option was not propagated to the generated
Makefile correctly:

```
$ poac b             
 Compiling <package name>
  Finished `dev` profile [unoptimized + debuginfo] target(s) in 4.83s

$ poac b -q
 Compiling <package name>
```

This is because Makefile is generated based on Poac's option and will
not reflect the option until it is regenerated.

This patch instead generates Makefile, which takes option info
dynamically. Also, this patch avoids printing `Compiling <package name>`
when the target is cached.