build: support lib.cc (#1035)
A simple implementation for static library building. It functions nearly
identically to executable building, except that we look for files named
`lib.cc` (or the other acceptable C++ extensions) and resolved the
dependencies for that file. Instead of linking everything into an
executable at the end, we instead use `ar` to build an object archive
called `lib<package-name>.a`.

---------

Co-authored-by: Krist Pregracke <krist@tiger-scm.com>
Co-authored-by: Ken Matsui <26405363+ken-matsui@users.noreply.github.com>